app/pyproject.toml
[tool.poetry]
name = "titier-backend"
version = "0.1.0"
description = "Backend do assistente de estudos PDF AI com IA local"
authors = ["matheusdanoite"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.128.5"
uvicorn = "^0.40.0"
llama-index = "^0.14.13"
pymupdf = "^1.26.7"
qdrant-client = "^1.16.2"
sentence-transformers = "^5.2.2"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.22"
huggingface-hub = "^1.4.1"
rapidocr-onnxruntime = "^1.4.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.28.0"

[tool.poetry.scripts]
serve = "app.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
